chore: updated getPerPage and getPage methods and added interface for CMS classes
