+++ AppAmbitSdk/Sources/Cms.swift
@@ -2,9 +2,9 @@ import Foundation
 
 @objcMembers
 public final class Cms: NSObject {
-    fileprivate nonisolated(unsafe) static var apiService: ApiService!
-    fileprivate nonisolated(unsafe) static var storageService: StorageService!
-    fileprivate static let fetchedContentTypes = ThreadSafeSet<String>()
+    internal nonisolated(unsafe) static var apiService: ApiService!
+    internal nonisolated(unsafe) static var storageService: StorageService!
+    internal static let fetchedContentTypes = ThreadSafeSet<String>()
 
     static func initialize(apiService: ApiService, storageService: StorageService) {
         self.apiService = apiService
@@ -29,285 +29,29 @@ public final class Cms: NSObject {
         return CmsQueryObjC(contentType: contentType)
     }
 
-    @objc(clearWithContentType:)
-    public static func clear(_ contentType: String) {
+    @objc(clearCacheWithContentType:)
+    public static func clearCache(_ contentType: String) {
         guard storageService != nil else { return }
         do {
             try storageService.deleteCmsData(contentType)
             fetchedContentTypes.remove(contentType)
         } catch {
-            debugPrint("Cms [clear error]: \(error)")
+            debugPrint("Cms [clearCache error]: \(error)")
         }
     }
 
     @objc
-    public static func clearAll() {
+    public static func clearAllCache() {
         guard storageService != nil else { return }
         do {
             try storageService.deleteAllCmsData()
             fetchedContentTypes.removeAll()
         } catch {
-            debugPrint("Cms [clearAll error]: \(error)")
+            debugPrint("Cms [clearAllCache error]: \(error)")
         }
     }
 }
 
-// MARK: - Generic Query Builder
-
-public final class CmsQuery<T: Decodable>: @unchecked Sendable {
-    private let contentType: String
-    private var whereClause: String = ""
-    private var args: [String] = []
-    private var orderBy: String?
-    private var page: Int?
-    private var perPage: Int?
-
-    init(contentType: String) {
-        self.contentType = contentType
-    }
-
-    private func addCondition(_ field: String, _ op: String, _ value: String) {
-        if !whereClause.isEmpty { whereClause += " AND " }
-        whereClause += "json_extract(value, '$.\(field)') \(op) ?"
-        args.append(value)
-    }
-
-    private func addNumericCondition(_ field: String, _ op: String, _ value: Any) {
-        if !whereClause.isEmpty { whereClause += " AND " }
-        whereClause += "CAST(json_extract(value, '$.\(field)') AS REAL) \(op) ?"
-        args.append("\(value)")
-    }
-
-    public func search(_ query: String) -> Self {
-        let trimmed = query.trimmingCharacters(in: .whitespacesAndNewlines)
-        if !trimmed.isEmpty {
-            if !whereClause.isEmpty { whereClause += " AND " }
-            whereClause += "value LIKE ?"
-            args.append("%\(trimmed)%")
-        }
-        return self
-    }
-
-    public func equals(_ field: String, _ value: String) -> Self { addCondition(field, "=", value); return self }
-    public func notEquals(_ field: String, _ value: String) -> Self { addCondition(field, "!=", value); return self }
-    public func contains(_ field: String, _ value: String) -> Self { addCondition(field, "LIKE", "%\(value)%"); return self }
-    public func startsWith(_ field: String, _ value: String) -> Self { addCondition(field, "LIKE", "\(value)%"); return self }
-    public func greaterThan(_ field: String, _ value: Any) -> Self { addNumericCondition(field, ">", value); return self }
-    public func greaterThanOrEqual(_ field: String, _ value: Any) -> Self { addNumericCondition(field, ">=", value); return self }
-    public func lessThan(_ field: String, _ value: Any) -> Self { addNumericCondition(field, "<", value); return self }
-    public func lessThanOrEqual(_ field: String, _ value: Any) -> Self { addNumericCondition(field, "<=", value); return self }
-
-    public func inList(_ field: String, _ values: [String]) -> Self {
-        if !whereClause.isEmpty { whereClause += " AND " }
-        let placeholders = Array(repeating: "?", count: values.count).joined(separator: ", ")
-        whereClause += "json_extract(value, '$.\(field)') IN (\(placeholders))"
-        args.append(contentsOf: values)
-        return self
-    }
-
-    public func notInList(_ field: String, _ values: [String]) -> Self {
-        if !whereClause.isEmpty { whereClause += " AND " }
-        let placeholders = Array(repeating: "?", count: values.count).joined(separator: ", ")
-        whereClause += "json_extract(value, '$.\(field)') NOT IN (\(placeholders))"
-        args.append(contentsOf: values)
-        return self
-    }
-
-    public func orderByAscending(_ field: String) -> Self {
-        self.orderBy = "json_extract(value, '$.\(field)') ASC"
-        return self
-    }
-
-    public func orderByDescending(_ field: String) -> Self {
-        self.orderBy = "json_extract(value, '$.\(field)') DESC"
-        return self
-    }
-
-    public func setPage(_ page: Int) -> Self { self.page = page; return self }
-    public func setPerPage(_ perPage: Int) -> Self { self.perPage = perPage; return self }
-
-    public func getList(completion: @escaping @Sendable ([T]) -> Void) {
-        let limit = perPage ?? -1
-        let offset = perPage != nil ? ((page ?? 1) - 1) * perPage! : 0
-
-        if Cms.fetchedContentTypes.contains(contentType) {
-            let cached = queryLocalCache(orderBy: orderBy, limit: limit, offset: offset)
-            completion(cached)
-            return
-        }
-
-        Cms.fetchedContentTypes.insert(contentType)
-
-        let cached = queryLocalCache(orderBy: orderBy, limit: limit, offset: offset)
-        if cached.isEmpty {
-            fetchAndReturn(orderBy: orderBy, limit: limit, offset: offset, completion: completion)
-        } else {
-            completion(cached)
-            refreshCacheInBackground()
-        }
-    }
-
-    private func queryLocalCache(orderBy: String?, limit: Int, offset: Int) -> [T] {
-        do {
-            let jsonList = try Cms.storageService.queryCmsData(
-                contentType: contentType,
-                whereClause: whereClause,
-                args: args,
-                orderBy: orderBy,
-                limit: limit,
-                offset: offset
-            )
-            let decoder = JSONDecoder()
-            return jsonList.compactMap { jsonString in
-                guard let data = jsonString.data(using: .utf8) else { return nil }
-                do {
-                    return try decoder.decode(T.self, from: data)
-                } catch {
-                    debugPrint("Cms [decode error] \(T.self): \(error)")
-                    return nil
-                }
-            }
-        } catch {
-            return []
-        }
-    }
-
-    private func fetchAndReturn(orderBy: String?, limit: Int, offset: Int, completion: @escaping @Sendable ([T]) -> Void) {
-        fetchAllRemoteDataSync { _ in
-            completion(self.queryLocalCache(orderBy: orderBy, limit: limit, offset: offset))
-        }
-    }
-
-    private func refreshCacheInBackground() {
-        fetchAllRemoteDataSync { _ in }
-    }
-
-    private func fetchAllRemoteDataSync(completion: @escaping @Sendable (Bool) -> Void) {
-        let perPageFetch = 20
-        let endpoint = CmsEndpoint(contentType: contentType, page: 1, perPage: perPageFetch)
-        
-        Cms.apiService.executeRequest(endpoint, responseType: [String: JSONValue].self) { result in
-            guard let responseDict = result.data else {
-                debugPrint("Cms [fetch error]: Unable to decode JSON or response is nil. Result error: \(String(describing: result.errorType))")
-                completion(false)
-                return
-            }
-
-            guard case let .array(dataArray) = responseDict["data"] else {
-                self.storeRawDict(responseDict, completion: completion)
-                return
-            }
-
-            var total = 0
-            if case let .object(metaObj) = responseDict["meta"],
-               case let .int(t) = metaObj["total"] {
-                total = t
-            }
-
-            let totalPages = Int(ceil(Double(total) / Double(perPageFetch)))
-
-            if totalPages <= 1 {
-                self.storeRawDict(responseDict, completion: completion)
-            } else {
-                self.fetchRemainingPages(startPage: 2, totalPages: totalPages, perPage: perPageFetch) { finalExtraItems in
-                    var allItems = dataArray
-                    allItems.append(contentsOf: finalExtraItems)
-                    var localDict = responseDict
-                    localDict["data"] = .array(allItems)
-                    self.storeRawDict(localDict, completion: completion)
-                }
-            }
-        }
-    }
-
-    private func fetchRemainingPages(startPage: Int, totalPages: Int, perPage: Int, completion: @escaping @Sendable ([JSONValue]) -> Void) {
-        let items = ThreadSafeArray<JSONValue>(initialItems: [])
-        let group = DispatchGroup()
-        
-        for p in startPage...totalPages {
-            group.enter()
-            let endpoint = CmsEndpoint(contentType: contentType, page: p, perPage: perPage)
-            Cms.apiService.executeRequest(endpoint, responseType: [String: JSONValue].self) { result in
-                if let dict = result.data, case let .array(nextData) = dict["data"] {
-                    items.append(contentsOf: nextData)
-                }
-                group.leave()
-            }
-        }
-        
-        group.notify(queue: .global()) {
-            completion(items.all())
-        }
-    }
-
-    private func storeRawDict(_ dict: [String: JSONValue], completion: @escaping @Sendable (Bool) -> Void) {
-        do {
-            let encoder = JSONEncoder()
-            let jsonData = try encoder.encode(dict)
-            if let jsonString = String(data: jsonData, encoding: .utf8) {
-                try Cms.storageService.putCmsData(contentType, jsonString)
-                debugPrint("Cms [store success]: Stored data for \(contentType)")
-                completion(true)
-            } else {
-                completion(false)
-            }
-        } catch {
-            debugPrint("Cms [store error]: Failed to store data for \(contentType). Error: \(error)")
-            completion(false)
-        }
-    }
-}
-
-// MARK: - ObjC Component
-
-@objcMembers
-public final class CmsQueryObjC: NSObject, @unchecked Sendable {
-    private let contentType: String
-    private let internalQuery: CmsQuery<JSONValue>
-
-    init(contentType: String) {
-        self.contentType = contentType
-        self.internalQuery = CmsQuery<JSONValue>(contentType: contentType)
-    }
-
-    @objc(search:)
-    public func search(_ query: String) -> Self { _ = internalQuery.search(query); return self }
-    @objc(equals:value:)
-    public func equals(_ field: String, _ value: String) -> Self { _ = internalQuery.equals(field, value); return self }
-    @objc(notEquals:value:)
-    public func notEquals(_ field: String, _ value: String) -> Self { _ = internalQuery.notEquals(field, value); return self }
-    @objc(contains:value:)
-    public func contains(_ field: String, _ value: String) -> Self { _ = internalQuery.contains(field, value); return self }
-    @objc(startsWith:value:)
-    public func startsWith(_ field: String, _ value: String) -> Self { _ = internalQuery.startsWith(field, value); return self }
-    @objc(greaterThan:value:)
-    public func greaterThan(_ field: String, _ value: Any) -> Self { _ = internalQuery.greaterThan(field, value); return self }
-    @objc(greaterThanOrEqual:value:)
-    public func greaterThanOrEqual(_ field: String, _ value: Any) -> Self { _ = internalQuery.greaterThanOrEqual(field, value); return self }
-    @objc(lessThan:value:)
-    public func lessThan(_ field: String, _ value: Any) -> Self { _ = internalQuery.lessThan(field, value); return self }
-    @objc(lessThanOrEqual:value:)
-    public func lessThanOrEqual(_ field: String, _ value: Any) -> Self { _ = internalQuery.lessThanOrEqual(field, value); return self }
-    @objc(inList:values:)
-    public func inList(_ field: String, _ values: [String]) -> Self { _ = internalQuery.inList(field, values); return self }
-    @objc(notInList:values:)
-    public func notInList(_ field: String, _ values: [String]) -> Self { _ = internalQuery.notInList(field, values); return self }
-    @objc(orderByAscending:)
-    public func orderByAscending(_ field: String) -> Self { _ = internalQuery.orderByAscending(field); return self }
-    @objc(orderByDescending:)
-    public func orderByDescending(_ field: String) -> Self { _ = internalQuery.orderByDescending(field); return self }
-    @objc(setPage:)
-    public func setPage(_ page: Int) -> Self { _ = internalQuery.setPage(page); return self }
-    @objc(setPerPage:)
-    public func setPerPage(_ perPage: Int) -> Self { _ = internalQuery.setPerPage(perPage); return self }
-
-    @objc(getListWithCompletion:)
-    public func getList(completion: @escaping @Sendable ([Any]) -> Void) {
-        internalQuery.getList { items in
-            completion(items.map { $0.toAny() })
-        }
-    }
-}
 
 // MARK: - JSON Handling
 
@@ -397,7 +141,7 @@ public struct AnyDecodable: Decodable, CustomStringConvertible, @unchecked Senda
     }
 }
 
-private final class ThreadSafeSet<T: Hashable>: @unchecked Sendable {
+internal final class ThreadSafeSet<T: Hashable>: @unchecked Sendable {
     private var set = Set<T>()
     private let lock = NSLock()
 
@@ -423,7 +167,7 @@ private final class ThreadSafeSet<T: Hashable>: @unchecked Sendable {
         set.removeAll()
     }
 }
-private final class ThreadSafeArray<T>: @unchecked Sendable {
+internal final class ThreadSafeArray<T>: @unchecked Sendable {
     private var array: [T]
     private let lock = NSLock()
